refactor: renames `Toster` file name
